split photons up instead of lines
Photons are much cheaper than lines so having extras on some
threads is less costly than whole lines. But it actually makes
more sense and simplifies the code by just rounding the number of
photons per line up such that all threads end up with the same
number to simulate.

As a side effect this means that numbers of lines less than the
core count can still be efficiently multithreaded.